Move slice definitions in slice.rs
